+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- Reduce spam in channel by reacting with thumbs up to message instead of sending new message to confirm
+- Update to only process messages in #general or DM from Arold for testing purposes
+
 ## 0.1.0
 
 - Initial development release